+++ packages/graph-explorer/src/core/featureFlags.ts
@@ -1,4 +1,4 @@
-import { atom, useAtomValue } from "jotai";
+import { atom, ExtractAtomValue, useAtomValue } from "jotai";
 import { atomWithLocalForage } from "./StateProvider/localForageEffect";
 
 /** Shows debug actions in various places around the app. */
@@ -13,17 +13,14 @@ export const allowLoggingDbQueryAtom = atomWithLocalForage(
   "allowLoggingDbQuery"
 );
 
-export type FeatureFlags = {
-  showDebugActions: boolean;
-  allowLoggingDbQuery: boolean;
-};
-
 export const featureFlagsSelector = atom(get => {
   return {
     showDebugActions: get(showDebugActionsAtom),
     allowLoggingDbQuery: get(allowLoggingDbQueryAtom),
-  } satisfies FeatureFlags;
+  };
 });
+
+export type FeatureFlags = ExtractAtomValue<typeof featureFlagsSelector>;
 
 export function useFeatureFlags() {
   return useAtomValue(featureFlagsSelector);